pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemimg"
version = "0.3.0"
description = "Lightweight wrapper for generating and editing images from Nano Banana / Gemini 2.5 Flash Image."
readme = "README.md"
license = "MIT"
authors = [{ name = "Max Woolf", email = "max@minimaxir.com" }]
keywords = ["gemini", "nano-banana", "image generation", "ai"]
classifiers = []
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "python-dotenv>=1.1.1",
    "orjson>=3.11.3",
    "Pillow>=11.3.0",
]

[project.urls]
Homepage = "https://github.com/minimaxir/gemimg"

[tool.setuptools.packages.find]
include = ["gemimg*"]
